planner: resolve cherry-pick conflicts for #70389
