+++ pkg/planner/core/operator/physicalop/BUILD.bazel
@@ -21,40 +21,3 @@ go_library(
         "@com_github_pingcap_tipb//go-tipb",
     ],
 )
-<<<<<<< HEAD
-=======
-
-go_test(
-    name = "physicalop_test",
-    timeout = "short",
-    srcs = [
-        "fragment_test.go",
-        "physical_batch_point_get_test.go",
-        "physical_utils_test.go",
-    ],
-    embed = [":physicalop"],
-    flaky = True,
-    shard_count = 6,
-    deps = [
-        "//pkg/expression",
-        "//pkg/expression/aggregation",
-        "//pkg/kv",
-        "//pkg/meta/model",
-        "//pkg/parser/ast",
-        "//pkg/parser/mysql",
-        "//pkg/planner/core/base",
-        "//pkg/planner/core/operator/logicalop",
-        "//pkg/planner/property",
-        "//pkg/types",
-        "//pkg/util/mock",
-        "@com_github_pingcap_tipb//go-tipb",
-        "@com_github_stretchr_testify//require",
-    ],
-)
-
-filegroup(
-    name = "generator_files",
-    srcs = ["plan_clone_generated.go"],
-    visibility = ["//visibility:public"],
-)
->>>>>>> 13fbbb2e3f0 (planner: optimize duplicate removal for common-handle batch point get (#70389))
